+++ package.json
@@ -15,8 +15,9 @@
   "scripts": {
     "start": "gobble",
     "build": "gobble build -f dist",
-    "prepublish": "npm run build",
+    "prepublish": "npm run lint && npm test",
     "test": "babel-node test/test.js",
+    "lint": "eslint src",
     "pretest": "npm run build && npm run create-viewer",
     "create-viewer": "babel-node test/create-viewer.js"
   },
