+++ src/utils/generateId.js
@@ -1,0 +1,3 @@
+export default function generateId () {
+	return 'ramjet' + ~~( Math.random() * 1000000 );
+}
