--- src/wrappers/wrapNode.js
@@ -1,5 +1,0 @@
-import HtmlWrapper from './HtmlWrapper';
-
-export default function wrapNode ( node, options ) {
-	return new HtmlWrapper( node, options );
-}
